requirements.txt
flask>=3.0.0
requests>=2.31.0
paramiko>=3.3.1
python-dotenv>=1.0.0
cryptography>=41.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
